refactor main_window.py into 7 focused mixin modules
Break up the 2037-line MainWindow god object (the #1 merge conflict
hotspot) into focused mixin classes, each in its own file:

- main_window_menus.py: menu bar construction, keybinding management
- main_window_file_ops.py: file I/O, clipboard, undo/redo, examples
- main_window_simulation.py: sim execution, result display, CSV export
- main_window_analysis.py: analysis type configuration dialogs
- main_window_view.py: theme, toggles, probe, zoom, image export
- main_window_print.py: print, print preview, PDF export
- main_window_settings.py: QSettings persistence, autosave, closeEvent

MainWindow inherits all mixins via Python MRO. All method names,
signatures, and import paths are preserved — zero API changes.

Additionally decomposes _display_simulation_results (254 lines) into
a dispatch table + 7 per-analysis-type handler methods.

main_window.py: 2037 → 327 lines (84% reduction)
New tests: 28 structural tests verifying mixin integrity

Closes #207
